all: use google.golang.org/protobuf directly
Now the only use of github.com/golang/protobuf is in the generated
db.pb.go file, which has an assertion that the old module is new enough
to be implemented in terms of the new module.

The new code is also simpler, as it doesn't error. When an input
timestamp is nil or zero, the APIs return a zero timestamp too.

While at it, remove references to the old gitlab.com/vocdoni/vocexplorer
module. All packages in the new github.com/vocdoni/vocexplorer module
should use imports from the same module, instead of imports from the old
and stale module. This was likely causing weird bugs and version skews.